factor out depresolve logic into separate pkg to make it callable from external packages
